+++ backend/canisters/community/impl/src/lifecycle/post_upgrade.rs
@@ -1,13 +1,13 @@
 use crate::jobs::import_groups::finalize_group_import;
-use crate::lifecycle::{init_env, init_state, UPGRADE_BUFFER_SIZE};
+use crate::lifecycle::{init_env, init_state};
 use crate::memory::get_upgrades_memory;
 use crate::{read_state, Data};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use community_canister::post_upgrade::Args;
 use ic_cdk_macros::post_upgrade;
-use ic_stable_structures::reader::{BufferedReader, Reader};
 use instruction_counts_log::InstructionCountFunctionId;
+use stable_memory::get_reader;
 use tracing::info;
 
 #[post_upgrade]
@@ -16,7 +16,7 @@ fn post_upgrade(args: Args) {
     let env = init_env();
 
     let memory = get_upgrades_memory();
-    let reader = BufferedReader::new(UPGRADE_BUFFER_SIZE, Reader::new(&memory, 0));
+    let reader = get_reader(&memory);
 
     let (data, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>) = serializer::deserialize(reader).unwrap();