+++ backend/canisters/user/impl/src/lifecycle/mod.rs
@@ -11,8 +11,6 @@ mod inspect_message;
 mod post_upgrade;
 mod pre_upgrade;
 
-const UPGRADE_BUFFER_SIZE: usize = 1024 * 1024; // 1MB
-
 fn init_env() -> Box<CanisterEnv> {
     ic_cdk_timers::set_timer(Duration::ZERO, reseed_rng);
     Box::default()